[Oztechan/CCC#1877] Move shared utils into :client:core:shared
