+++ android/feature/mobile/src/main/kotlin/com/oztechan/ccc/android/feature/mobile/ui/settings/SettingsFragment.kt
@@ -28,7 +28,7 @@ import com.oztechan.ccc.android.feature.mobile.util.showSnack
 import com.oztechan.ccc.android.feature.mobile.util.visibleIf
 import com.oztechan.ccc.client.core.analytics.AnalyticsManager
 import com.oztechan.ccc.client.core.analytics.model.ScreenName
-import com.oztechan.ccc.client.core.viewmodel.util.MAXIMUM_FLOATING_POINT
+import com.oztechan.ccc.client.core.shared.util.MAXIMUM_FLOATING_POINT
 import com.oztechan.ccc.client.model.AppTheme
 import com.oztechan.ccc.client.util.numberToIndex
 import com.oztechan.ccc.client.viewmodel.settings.SettingsEffect
