+++ client/core/shared/src/androidMain/kotlin/com/oztechan/ccc/client/core/shared/util/AndroidFormatUtil.kt
@@ -1,4 +1,4 @@
-package com.oztechan.ccc.client.core.viewmodel.util
+package com.oztechan.ccc.client.core.shared.util
 
 import java.text.DecimalFormat
 import java.text.DecimalFormatSymbols
