+++ client/core/shared/src/commonMain/kotlin/com/oztechan/ccc/client/core/shared/util/FormatUtil.kt
@@ -1,4 +1,4 @@
-package com.oztechan.ccc.client.core.viewmodel.util
+package com.oztechan.ccc.client.core.shared.util
 
 const val MAXIMUM_FLOATING_POINT = 9
