+++ client/core/shared/src/iosMain/kotlin/com/oztechan/ccc/client/core/shared/util/IOSFormatUtil.kt
@@ -1,4 +1,4 @@
-package com.oztechan.ccc.client.core.viewmodel.util
+package com.oztechan.ccc.client.core.shared.util
 
 import platform.Foundation.NSNumber
 import platform.Foundation.NSNumberFormatter
